+++ packages/types/CHANGELOG.md
@@ -1,5 +1,7 @@
 # @flatkit/types
 
+## 0.10.0
+
 ## 0.9.0
 
 ### Minor Changes